Move humidex function and tests to distinct modules
