+++ tests/hi_test.py
@@ -1,7 +1,7 @@
 # coding utf-8
 import pytest
 
-from ladybug_comfort.hi import heat_index, heat_index_warning_category, humidex
+from ladybug_comfort.hi import heat_index, heat_index_warning_category
 
 from ladybug.epw import EPW
 from ladybug.datacollection import HourlyContinuousCollection
@@ -40,12 +40,3 @@ def test_heating_degree_time_collection():
     assert isinstance(hourly_category, HourlyContinuousCollection)
     assert len(hourly_category.values) == calc_length
     assert hourly_category[4000] == 1
-
-
-def test_humidex():
-    """Test the humidex function."""
-    assert humidex(22, 15) == pytest.approx(25.96900, rel=1e-3)
-    assert humidex(20, 20) == pytest.approx(27.56977, rel=1e-3)
-    assert humidex(28, 20) == pytest.approx(35.56977, rel=1e-3)
-    assert humidex(28, 26) == pytest.approx(41.45899, rel=1e-3)
-    assert humidex(34, 25) == pytest.approx(46.33840, rel=1e-3)
